apps/continuous_mfa/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "continuous_mfa"
version = "0.1.0"
description = "A FastAPI application in a monorepo"
authors = [{ name = "GroG Bot", email = "grogbot@robotlab-x.com" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "python-multipart",
    "PyJWT"
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.pytest.ini_options]
testpaths = ["tests"]